+++ webui/module/Director/src/Director/Controller/DirectorController.php
@@ -5,7 +5,7 @@
  * bareos-webui - Bareos Web-Frontend
  *
  * @link      https://github.com/bareos/bareos for the canonical source repository
- * @copyright Copyright (C) 2013-2024 Bareos GmbH & Co. KG (http://www.bareos.org/)
+ * @copyright Copyright (C) 2013-2025 Bareos GmbH & Co. KG (http://www.bareos.org/)
  * @license   GNU Affero General Public License (http://www.gnu.org/licenses/)
  *
  * This program is free software: you can redistribute it and/or modify
@@ -170,11 +170,17 @@ class DirectorController extends AbstractActionController
             );
         }
 
+        $reportDetail = "";
+
+        if ($this->params()->fromQuery('detail') == true) {
+            $reportDetail = "all";
+        }
+
         $result = null;
 
         try {
             $this->bsock = $this->getServiceLocator()->get('director');
-            $result = $this->getDirectorModel()->getBackupUnitReport($this->bsock, 2, "all");
+            $result = $this->getDirectorModel()->getBackupUnitReport($this->bsock, 2, $reportDetail);
             $this->bsock->disconnect();
         } catch (Exception $e) {
             echo $e->getMessage();